Add further refactoring
