+++ src/components/App/misc/Forgot.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const Forgot = () => (
+  <div className="Forgot">
+    Forgot Page
+  </div>
+);
+
+export default Forgot;
